requirements.txt
fastapi==0.101.0
sqlalchemy==2.0.19
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
httpagentparser==1.9.5
python-multipart==0.0.6
requests==2.31.0
